requirement.txt
numpy==1.14.2
opencv-python==3.4.2.16
pandas==0.23.1
Pillow==5.2.0
pyparsing==2.2.0
tensorflow==1.8.0
tqdm==4.23.4